Move scriptable objects to relevant folders
