--- Assets/00_Content/Scripts/Interactables/BeerType.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Interactables {
-	public enum BeerType {
-		Lager,
-		Stout,
-		Ale
-	}
-}
