+++ Assets/00_Content/Scripts/Interactables/Beers.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dd1de59831c7407e9dadeabcde8933b5
+timeCreated: 1618559238
