+++ Assets/00_Content/Scripts/Interactables/Beers/BeerData.cs.meta
