+++ Assets/00_Content/Scripts/Interactables/Beers/BeerTap.cs
@@ -1,10 +1,6 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-
-namespace Interactables {
+namespace Interactables.Beers {
 
 	public class BeerTap : Interactable {
