+++ Assets/00_Content/Scripts/Interactables/Beers/BeerType.cs
@@ -1,0 +1,7 @@
+﻿namespace Interactables.Beers {
+	public enum BeerType {
+		Lager,
+		Stout,
+		Ale
+	}
+}
