--- Assets/00_Content/Scripts/ScriptableObjects.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c4873870e4334ab38368c05d36be6f28
-timeCreated: 1618306410
