+++ Assets/00_Content/Scripts/Vikings/VikingData.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace ScriptableObjects {
+namespace Vikings {
 	[CreateAssetMenu(fileName = "new Viking", menuName = "Game/VikingData", order = 0)]
 	public class VikingData : ScriptableObject {
 		[Range(0, 100)]
