+++ Assets/00_Content/Scripts/Vikings/VikingStats.cs
@@ -1,5 +1,4 @@
 ﻿using Rounds;
-using ScriptableObjects;
 using UnityEngine;
 
 namespace Vikings {
